+++ packages/eyes-api/src/output/TestResultContainer.ts
@@ -1,24 +1,24 @@
 import * as utils from '@applitools/utils'
-import {TestResults, TestResultsData} from './TestResults'
+import type * as types from '@applitools/types'
+import {TestResults, TestResultsData, DeleteTestFunc} from './TestResults'
 
-export type TestResultContainer = {
-  readonly exception: Error
-  readonly testResults: TestResults
-}
+export type TestResultContainer = types.TestResultContainer
 
-export class TestResultContainerData implements Required<TestResultContainer> {
-  private _container: TestResultContainer = {} as any
+export class TestResultContainerData implements Required<types.TestResultContainer> {
+  private _container: types.TestResultContainer = {} as any
+  private _deleteTest: DeleteTestFunc
 
   /** @internal */
-  constructor(container: TestResultContainer) {
+  constructor(container: types.TestResultContainer, deleteTest: DeleteTestFunc) {
     this._container = container instanceof TestResultContainerData ? container.toJSON() : container
+    this._deleteTest = deleteTest
   }
 
   get testResults(): TestResults {
     return this._container.testResults
   }
   getTestResults(): TestResultsData {
-    return new TestResultsData(this.testResults)
+    return new TestResultsData(this.testResults, this._deleteTest)
   }
 
   get exception(): Error {
@@ -28,13 +28,21 @@ export class TestResultContainerData implements Required<TestResultContainer> {
     return this.exception
   }
 
+  get browserInfo(): types.BrowserInfoRenderer {
+    return this._container.browserInfo
+  }
+
+  getBrowserInfo(): types.BrowserInfoRenderer {
+    return this.browserInfo
+  }
+
   /** @internal */
-  toObject(): TestResultContainer {
+  toObject(): types.TestResultContainer {
     return this._container
   }
 
   /** @internal */
-  toJSON(): TestResultContainer {
+  toJSON(): types.TestResultContainer {
     return utils.general.toJSON(this._container)
   }