+++ packages/eyes-playwright-universal/src/universal-client.ts
@@ -194,8 +194,8 @@ export class EyesManager implements types.EyesManager<Driver, Element, Selector>
     return new Eyes({eyes, socket: this._socket, refer: this._refer})
   }
 
-  async closeAllEyes({throwErr}: {throwErr?: boolean} = {}): Promise<types.TestResult[]> {
-    return this._socket.request('EyesManager.closeAllEyes', {manager: this._manager, throwErr})
+  async closeManager({throwErr}: {throwErr?: boolean} = {}): Promise<types.TestResultSummary[]> {
+    return this._socket.request('EyesManager.closeManager', {manager: this._manager, throwErr})
   }
 }