+++ packages/eyes-api/CHANGELOG.md
@@ -1,8 +1,8 @@
 # Changelog
 
 ## Unreleased
-- updated to @applitools/logger@1.0.10 (from 1.0.9)
 
+- change EyesManager.closeAllEyes to EyesManager.closeManager 
 
 ## 1.1.8 - 2022/2/16